Update meter so value is set to pos along curve
When the user drags along the gauge, the needle doesn't point to exactly
where the user has dragged. This is especially when the user is trying
to set the value right at the boundary of a colored segment by clicking
on the curve.

This translates the user's mouse value to be the angle along the curve
of the gauge, instead of just the linear distance across the gauge.

Now if the user drags to a point along the curve, the needle will point
there.