Fix lid state-setting so controls rerender
Previously, the container lid would update, but React wouldn't recognize
the state as having changed, because the only property changing was 
nested, so the container button would not update until the next render
(clicking something else).

By copying `containerLid` into a new object, we force React to recognize
the state change and rerender the button.